feat(core): use 64-bit app root timestamps
[no changelog]

+++ core/embed/sys/syscall/stm32/syscall_dispatch.c
@@ -1064,7 +1064,7 @@ __attribute((no_stack_protector)) void syscall_handler(uint32_t *args,
 
     case SYSCALL_APP_ROOT_GET_TIMESTAMP: {
       app_ring_t ring = (app_ring_t)args[0];
-      uint32_t *timestamp = (uint32_t *)args[1];
+      int64_t *timestamp = (int64_t *)args[1];
       ts_t status = app_root_get_timestamp__verified(ring, timestamp);
       args[0] = ts_code(status);
     } break;
